search page cleanup cont. - move entity info query code to a seperate file - minor search css changes
