+++ src/styles/features/search.css
@@ -9,7 +9,7 @@
   box-sizing: border-box;
   padding: 10px;
   padding-bottom: 0;
-  min-height: 12vh;
+  border-bottom: 1px solid var(--light-base-colour-dark);
 }
 
 .search-header {
@@ -68,12 +68,12 @@
   display: flex;
   align-content: space-around;
   background-color: white;
-  box-shadow: 0 3px 6px rgba(0,0,0,0.16), 0 3px 6px rgba(0,0,0,0.23);
+  box-shadow: 0 2px 3px rgba(0,0,0,0.1), 0 3px 6px rgba(0,0,0,0.2);
   transition: box-shadow 0.15s ease-in-out;
 }
 
 .search-searchbar:hover {
-  box-shadow: 0 4px 8px rgba(0,0,0,0.19), 0 6px 6px rgba(0,0,0,0.23);
+  box-shadow: 0 2px 4px rgba(0,0,0,0.1), 0 6px 6px rgba(0,0,0,0.2);
 }
 .search-searchbar > input {
   background: none;
@@ -96,7 +96,7 @@
 }
 
 .search-suggestions {
-  padding-top: 10px;
+  padding: 10px 0;
   font-size: 0.75em;
   color: var(--dark-base-colour);
 }
